WIP: refactor haplotype functions for Af1
